Improved string-finding efficiency
ASCII characters are determined by calling isprint() with the locale set to C, Latin-1
with the locale set to iso88591. The locale was set inside is_ascii() and is_latin()
functions passed to read_string(). Changed so the locale is set in inspect() and
read_string() simply calls isprint().

Updated tests to distinguish ASCII strings from those with Latin-1 characters.