Express ranges as strings readable by the target type
Double and float ranges are now expressed as numbers rather than as exponents. I.e. as 1e3
instead of 3.

Added a minimum to the range spec

The range spec can now be low:high or low:high:min. If specified, values within min of
zero (positive or negative) but not exactly zero are excluded. E.g. for doubles with a
range spec of -1e6:1e6:1e-6, 2e-8 or -5e7 would not be shown, but exactly zero would.

Bail out if the stream goes bad while consuming characters of a too-long string. This
avoids hanging the program if we hit EOF.